requirements.txt
# Core dependencies
openai-whisper>=20230314
pyaudio>=0.2.13
numpy>=1.20.0
soundfile>=0.12.1
pydub>=0.25.1
torch>=1.10.0
ffmpeg-python>=0.2.0

# UI dependencies
pyqt6>=6.2.0

# Utility dependencies
click>=8.0.0
python-dotenv>=0.19.0

# Testing dependencies
pytest>=7.0.0
